:fire: prune: Adjust program/session templates to handle archived editions
