+++ gatsby/createProgramSessionPages.js
@@ -26,6 +26,13 @@ const allowedTypes = [
   // 'en-direct',
 ];
 
+const planningFields = `
+  beginsAt
+  id
+  title
+  type
+`;
+
 /**
  * This function queries Gatsby's GraphQL server and asks for
  * All Plannings from Swapcard. If there are any GraphQL error it throws an error
@@ -38,42 +45,34 @@ const getPlannings = async ({ graphql, reporter, variables }) => {
   const { eventIds, page, pageSize } = variables;
   const result = await graphql(`
     {
-      edition2023: swapcard {
+      edition2024: swapcard {
         plannings(
-          eventId: "${eventIds[2023]}"
+          eventId: "${eventIds[2024]}"
           page: ${page}
           pageSize: ${pageSize}
         ) {
-          beginsAt
-          id
-          title
-          type
+          ${planningFields}
         }
       }
-
-      edition2022: swapcard {
-        plannings(
-          eventId: "${eventIds[2022]}"
-          page: ${page}
-          pageSize: ${pageSize}
-        ) {
-          beginsAt
-          id
-          title
-          type
+      edition2023: allProgram2023Json {
+        plannings: edges {
+          node {
+            ${planningFields}
+          }
         }
       }
-
-      edition2021: swapcard {
-        plannings(
-          eventId: "${eventIds[2021]}"
-          page: ${page}
-          pageSize: ${pageSize}
-        ) {
-          beginsAt
-          id
-          title
-          type
+      edition2022: allProgram2022Json {
+        plannings: edges {
+          node {
+            ${planningFields}
+          }
+        }
+      }
+      edition2021: allProgram2021Json {
+         plannings: edges {
+          node {
+            ${planningFields}
+          }
         }
       }
     }
@@ -87,16 +86,15 @@ const getPlannings = async ({ graphql, reporter, variables }) => {
     return;
   }
 
-  const {
-    data: { edition2021, edition2022, edition2023 },
-  } = result;
+  const { edition2024, edition2023, edition2022, edition2021 } = result.data;
 
-  const plannings = [
+  const previousEditions = [
+    ...edition2023.plannings,
     ...edition2022.plannings,
     ...edition2021.plannings,
-    ...edition2023.plannings,
-  ];
-  // const plannings = { edition2022, edition2021 };
+  ].map((p) => p.node);
+
+  const plannings = [...edition2024.plannings, ...previousEditions];
 
   // eslint-disable-next-line consistent-return
   return plannings;
@@ -111,6 +109,9 @@ const createSession = async ({
 }) => {
   const { createPage } = actions;
   const template = path.resolve('src/templates/Program/Session/Session.jsx');
+  const archivedTemplate = path.resolve(
+    'src/templates/Program/Session/ArchivedSession.jsx'
+  );
 
   const { eventIds, page, pageSize } = variables;
 
@@ -141,18 +142,31 @@ const createSession = async ({
 
       reporter.info(sessionPath);
 
+      const context = {
+        planningIds: [id],
+        pageNumber: pageIndex + 1,
+        isLastPage: eventYear === 2021 && eventDates.length === pageIndex + 1,
+      };
+
+      if (eventYear === 2024) {
+        createPage({
+          path: sessionPath,
+          component: template,
+          context: {
+            eventId: eventIds[eventYear],
+            page,
+            pageSize,
+            ...context,
+          },
+        });
+
+        return;
+      }
+
       createPage({
         path: sessionPath,
-        component: template,
-        context: {
-          eventId: eventIds[eventYear],
-          page,
-          pageSize,
-          planningIds: [id],
-          pageNumber: pageIndex + 1,
-          isLastPage:
-            eventYear === '2021' && eventDates.length === pageIndex + 1,
-        },
+        component: archivedTemplate,
+        context,
       });
     });
 };
@@ -166,13 +180,16 @@ const createProgram = async ({
 }) => {
   const { createPage } = actions;
   const template = path.resolve('src/templates/Program/Program.jsx');
+  const archivedTemplate = path.resolve(
+    'src/templates/Program/ArchivedProgram.jsx'
+  );
 
   const totalPages = dates.length;
 
   reporter.info('creating Program pages:');
 
   Object.entries(dates).forEach(async (year) => {
-    const eventYear = year[0];
+    const eventYear = Number(year[0]);
     const eventDates = year[1];
 
     eventDates.forEach(async (date, index, array) => {
@@ -197,28 +214,40 @@ const createProgram = async ({
 
       const pagePaths = array.map((_, i) => getPagePath(i + 1));
 
-      const { eventIds, page, pageSize } = variables;
-      // createPage is an action passed to createPages
-      // See https://www.gatsbyjs.com/docs/actions#createPage for more info
+      const context = {
+        pageNumber,
+        planningIds,
+        nextPagePath: getPagePath(pageNumber + 1),
+        previousPagePath: getPagePath(pageNumber - 1),
+        pagePaths,
+        eventDates,
+      };
+
+      if (eventYear === 2024) {
+        const { eventIds, page, pageSize } = variables;
+        // createPage is an action passed to createPages
+        // See https://www.gatsbyjs.com/docs/actions#createPage for more info
+        createPage({
+          path: getPagePath(pageNumber),
+          // use the schedule template as the page component
+          component: template,
+          // `context` is available in the template as a prop and
+          // as a variable in GraphQL.
+          context: {
+            eventId: eventIds[eventYear],
+            page,
+            pageSize,
+            ...context,
+          },
+        });
+
+        return;
+      }
+
       createPage({
         path: getPagePath(pageNumber),
-        // use the schedule template as the page component
-        component: template,
-
-        // `context` is available in the template as a prop and
-        // as a variable in GraphQL.
-        context: {
-          eventId: eventIds[eventYear],
-          page,
-          pageSize,
-          pageNumber,
-          planningIds,
-          nextPagePath: getPagePath(pageNumber + 1),
-          previousPagePath: getPagePath(pageNumber - 1),
-          pagePaths,
-          // eventYear,
-          eventDates,
-        },
+        component: archivedTemplate,
+        context,
       });
     });
   });
@@ -227,12 +256,10 @@ const createProgram = async ({
 module.exports = async ({ graphql, actions, reporter }) => {
   const variables = {
     eventIds: {
-      2023: `${process.env.SWAPCARD_EVENT_ID}`,
-      2022: `${process.env.SWAPCARD_EVENT_ID_2022}`,
-      2021: `${process.env.SWAPCARD_EVENT_ID_2021}`,
+      2024: `${process.env.SWAPCARD_EVENT_ID}`,
     },
     page: 1,
-    pageSize: 250,
+    pageSize: 300,
   };
 
   reporter.info('start fetching data from Swapcard');
@@ -272,7 +299,7 @@ module.exports = async ({ graphql, actions, reporter }) => {
   /**
    * FIXME: Remove last date from 2021 edition temporarily until fixing it
    */
-  datesGroupByYear[2021].pop();
+  datesGroupByYear[2021]?.pop();
 
   await createSession({
     plannings,
