+++ src/templates/Program/Session/index.js
@@ -1,1 +1,2 @@
-export { default } from './Session';
+export { default as Session } from './Session';
+export { default as ArchivedSession } from './ArchivedSession';
