+++ src/templates/Program/index.js
@@ -1,1 +1,2 @@
-export { default } from './Program';
+export { default as Program } from './Program';
+export { default as ArchivedProgram } from './ArchivedProgram';
